requirements.txt
torch
torchvision
timm
numpy
pandas
scikit-learn
matplotlib
opencv-python
Pillow
tqdm
python-osc
